# docs/swagger.yaml
basePath: /
definitions:
  archive.DeviceSummary:
    properties:
      hostname:
        type: string
      id:
        type: string
      lastIngestedAt:
        type: string
      name:
        type: string
      rawBytes:
        type: integer
      sessions:
        type: integer
    type: object
...
  archive.IntegrityIssue:
    properties:
      path:
        type: string
      problem:
        type: string
      sessionId:
        type: string
    type: object
  archive.IntegrityResult:
    properties:
      checked:
        type: integer
      issues:
        items:
          $ref: '#/definitions/archive.IntegrityIssue'
        type: array
      missingFile:
        type: integer
      missingPath:
        type: integer
      missingSha:
        type: integer
      ok:
        type: integer
      sizeMismatch:
        type: integer
    type: object
  archive.RepositorySummary:
    properties:
      id:
        type: string
      lastIngestedAt:
        type: string
      name:
        type: string
      rawBytes:
        type: integer
      sessions:
        type: integer
      url:
        type: string
    type: object
  archive.Status:
    properties:
      devices:
        type: integer
      messages:
        type: integer
      missingRawFiles:
        type: integer
      missingRawSha:
        type: integer
      newestIngestedAt:
        type: string
      newestSessionTime:
        type: string
      oldestIngestedAt:
        type: string
      oldestSessionTime:
        type: string
      rawBytes:
        type: integer
      sessionEvents:
        type: integer
      sessions:
        type: integer
      toolEvents:
        type: integer
      usageEvents:
        type: integer
    type: object
...
  httpapi.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  httpapi.MeResponse:
    properties:
      role:
        type: string
    type: object
  httpapi.StatusResponse:
    properties:
      status:
        type: string
    type: object
...
  projects.ProjectSummary:
    properties:
      costUsd:
        type: number
      cwd:
        type: string
      displayName:
        type: string
      id:
        type: string
      relativePath:
        type: string
      repository:
        type: string
      repositoryUrl:
        type: string
      sessions:
        type: integer
      totalTokens:
        type: integer
    type: object
  projects.RepositorySummary:
    properties:
      costUsd:
        type: number
      id:
        type: string
      name:
        type: string
      owner:
        type: string
      repositoryUrl:
        type: string
      sessions:
        type: integer
      totalTokens:
        type: integer
    type: object
...
info:
  contact: {}
  description: API for Codex usage ingestion, session browsing, usage reporting, and
    archive inspection.
  title: codex-usage API
  version: 0.1.0
paths:
  /api/archive/devices:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/archive.DeviceSummary'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Archive devices
      tags:
      - Archive
...
  /api/archive/integrity:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/archive.IntegrityResult'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Archive integrity
      tags:
      - Archive
  /api/archive/repositories:
    get:
      parameters:
      - description: Maximum number of repositories
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/archive.RepositorySummary'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Archive repositories
      tags:
      - Archive
  /api/archive/status:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/archive.Status'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Archive status
      tags:
      - Archive
  /api/auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Login request
        in: body
        name: request
        required: true
        schema:
          properties:
            password:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httpapi.StatusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
      summary: Login
      tags:
      - Auth
...
  /api/auth/me:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httpapi.MeResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Current user
      tags:
      - Auth
...
  /api/projects:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/projects.ProjectSummary'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
      security:
      - CookieAuth: []
      summary: List projects
      tags:
      - Projects
  /api/repositories:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/projects.RepositorySummary'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
      security:
      - CookieAuth: []
      summary: List repositories
      tags:
      - Projects
...
securityDefinitions:
  CookieAuth:
    description: Admin session cookie. Login first, then send the codex_usage_session
      cookie.
    in: header
    name: Cookie
    type: apiKey
  DeviceTokenAuth:
    description: 'Device ingest token. Authorization: Bearer <token> is also accepted
      by the server.'
    in: header
    name: X-Device-Token
    type: apiKey
swagger: "2.0"